requirements.txt
psycopg2==2.8.5
Flask-SQLAlchemy==2.4.3
Flask==1.1.2
Flask-Migrate==2.5.3
python-slugify==4.0.0
requests==2.24.0
